+++ lua/lazy/health.lua
@@ -30,6 +30,15 @@ function M.check()
     vim.health.report_ok("no existing packages found by other package managers")
   end
 
+  for _, name in ipairs({ "packer", "plugged", "paq" }) do
+    for _, path in ipairs(vim.opt.rtp:get()) do
+      if path:find(name, 1, true) then
+        vim.health.report_error("Found paths on the rtp from another plugin manager `" .. name .. "`")
+        break
+      end
+    end
+  end
+
   local packer_compiled = vim.fn.stdpath("config") .. "/plugin/packer_compiled.lua"
   if vim.loop.fs_stat(packer_compiled) then
     vim.health.report_error("please remove the file `" .. packer_compiled .. "`")
@@ -75,7 +84,7 @@ function M.check_override(plugin)
   end
 
   local Handler = require("lazy.core.handler")
-  local skip = { "dependencies", "_", "opts" }
+  local skip = { "dependencies", "_", "opts", 1 }
   vim.list_extend(skip, vim.tbl_values(Handler.types))
 
   for key, value in pairs(plugin._.super) do
@@ -94,6 +103,7 @@ M.valid = {
   "commit",
   "cond",
   "config",
+  "deactivate",
   "dependencies",
   "dev",
   "dir",
@@ -104,6 +114,7 @@ M.valid = {
   "init",
   "keys",
   "lazy",
+  "main",
   "module",
   "name",
   "opts",