+++ lua/lazy/init.lua
@@ -33,15 +33,22 @@ function M.setup(spec, opts)
   end
   local start = vim.loop.hrtime()
 
-  -- load module cache before anything else
+  -- use the NEovim cache if available
+  if vim.loader then
+    package.loaded["lazy.core.cache"] = vim.loader
+  end
+
+  local Cache = require("lazy.core.cache")
+
   local enable_cache = not (
     opts
     and opts.performance
     and opts.performance.cache
     and opts.performance.cache.enabled == false
   )
+  -- load module cache before anything else
   if enable_cache then
-    require("lazy.core.cache").enable()
+    Cache.enable()
   end
 
   require("lazy.stats").track("LazyStart")
@@ -53,7 +60,7 @@ function M.setup(spec, opts)
   table.insert(package.loaders, 3, Loader.loader)
 
   if vim.g.profile_loaders then
-    require("lazy.core.cache").profile_loaders()
+    Cache.profile_loaders()
   end
 
   Util.track({ plugin = "lazy.nvim" }) -- setup start